requirements.txt
todoist-python >= 8.1.1
robotframework >= 3.2.1
robotframework-seleniumlibrary >= 4.4.0
robotframework-pageobjectlibrary >= 1.0.2
webdrivermanager >= 0.8.0
